requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
sentence_transformers
faiss-cpu
langchain-community
